Make user.addTask() work using extra arrow funct.
Before, we were exporting addTodo as:
- { addTodo: directTodos.addTodo }, which means we
were exporting the direct reference to addTodo object.
- But exporting this means that access to the
context (directTodos) is lost,
- So, calling user.addTodo() wasn't recognizing the
function at all.

Solution: use an extra function that executes
directTodos.addTodo() and export that instead. That way,
we won't have to worry about context, the function is
being executed inside the module, so it has access to
directTodos.