Solve project page add and delete anamolies
Before this, the adding of new project task made addition
of todos happen on all the projects. And same with deleting.
When todo of id 1 was deleted, todos having id 1 in all projects
were deleted.
That happened due to addevent listener targeting not just 1
but all projects.
Now this is solved, by getting project id from todo itself
when configuring the buttons instead from different source.

Also, the configure todo button configured todo for each
todo, this worked, but also was redundant.
Solve this.